feat:ros的数据删除时也需要同步到官网上

+++ ses-app/ses-web-ros/src/main/java/com/redescooter/ses/web/ros/service/restproduction/impl/SaleScooterServiceImpl.java
@@ -151,8 +151,20 @@ public class SaleScooterServiceImpl implements SaleScooterService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public GeneralResult deleteSaleScooter(IdEnter enter) {
+        OpeSaleScooter saleScooter = opeSaleScooterService.getById(enter.getId());
+        if (saleScooter == null) {
+            throw new SesWebRosException(ExceptionCodeEnums.PRODUCT_IS_NOT_EXIST.getCode(), ExceptionCodeEnums.PRODUCT_IS_NOT_EXIST.getMessage());
+        }
         opeSaleScooterService.removeById(enter.getId());
+        // 删除销售车辆的时候  需要把官网的数据也删除掉
+        syncDeleteData(saleScooter.getProductCode());
         return new GeneralResult(enter.getRequestId());
+    }
+
+
+    @Async
+    void syncDeleteData(String productionCode){
+        productionService.syncDeleteData(productionCode);
     }
